move all constants to consts file
