--- src/backend/api/consts.ts
@@ -1,7 +1,0 @@
-export const server_port = 4000
-export const log_sheet_name = "Log"
-export const loggedin_sheet_name = "Logged In"
-// export const hours_spreadsheet_id = '10AcplDpfbXlECQYaFuTFcIeN2U8raP9XysuN3e31js0' // REAL SHEET
-export const hours_spreadsheet_id = '1KxS5gqUhygWnLlga9wM2k6jfJr5tCKCxycN3axkWnCA' // TEST SHEET
-
-export const logPath = './api/memberlog.jsonl'
