+++ src/backend/api/memberlog.ts
@@ -1,6 +1,6 @@
 import fs from 'fs'
 import { LoggedIn } from '.'
-import { logPath } from './consts';
+import { memberLogFilePath } from '../../consts';
 import { updateLoggedIn } from './spreadsheet'
 
 
@@ -23,7 +23,7 @@ export async function logMember(name:string, loggingin:boolean, logged_in:Logged
 export function saveMemberLog() {
     // Use jsonl to avoid having to load the entire file into memory
     memberlog.forEach(item => {
-        fs.appendFileSync(logPath, JSON.stringify(item))
+        fs.appendFileSync(memberLogFilePath, JSON.stringify(item))
     })
     setMemberlog([])
 }
